LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-canvasImageSource-expected.txt



FAIL <video> and VideoFrame constructed VideoFrame assert_unreached: Reached unreachable code
FAIL CSSImageValue constructed VideoFrame assert_throws_dom: CSSImageValues are currently always tainted function "_ => { new VideoFrame(bgImage, {timestamp: 0}); }" did not throw
FAIL Image element constructed VideoFrame assert_equals: top left corner expected "Yellow" but got "#0"
FAIL SVGImageElement constructed VideoFrame Type error
FAIL Canvas element constructed VideoFrame assert_equals: top left corner expected "Yellow" but got "#0"
FAIL Copy of canvas element constructed VideoFrame assert_equals: top left corner expected "Yellow" but got "#0"
